Initial commit of Retrospective Timesheet
The idea behind this project is to help analyze usage of a system to try
to determine when the system was being used, and for what.

The farther back this code goes, the less reliable it will be, but
hopefully it will be helpful for getting people out of a pinch if they
need to figure out the amount of time they have been working on a
particular project.