Use correct Chrome time, and bypass implicit hooks
Not sure why .create fails and .create! succeeds for what appears to be
valid data when inspecting by hand... The only difference between these
methods is that the "!" method does not run hooks, so since I did not
declare any hooks, that means this must be the result of an implicit
hook... (Hence the commit message)

I will have to look into this later.