Fleshed out and moved main Gui logic. (Not appearance)
Now the Gui class is simply a dispatcher for other GUIs, and will
contain the logic of when to show and hide them.

MainWindow will be the class that holds onto an instance of the root
window, so it will be easy to find, and to kill the Tk runtime when the
window closes.